requirements.txt
openai-whisper
pytube
requests
streamlit
pydub
numpy